Refactor acceptance tests to use common user-group-server
